test: add test for checking sysfs integrity
Commit 4dde420fc3e2 ("mdadm: remove container_enough logic") introduced
a bug, where array sysfs intro is obtained from first drive added and
never updated. This results in info mismatch if not "up to date" drive
is added first. Due to this issue, the commit was reverted.

Add test for checking if sysfs entries are correctly set according to
metadata.

Check if consistency_policy: ppl is respected.

Signed-off-by: Mateusz Kusiak <mateusz.kusiak@intel.com>