mdadm: deprecate bitmap custom file
This option has been deprecated in kernel by Christoph in commit
0ae1c9d38426 ("md: deprecate bitmap file support"). Do the same in
mdadm.

With this change, user must acknowledge it, it is not
skippable. The implementation of custom bitmap file looks like it's
abandoned. It cannot be done by Incremental so it is not respected by
any udev based system and it seems to not be recorded by metadata.
User must assemble such volume manually.

Tests for bitmap custom file are removed because now they will not
pass because interaction with user is mandatory.

Signed-off-by: Mariusz Tkaczyk <mariusz.tkaczyk@linux.intel.com>