imsm: make freesize required to volume autolayout
Autolayout_imsm() shall be executed when IMSM_NO_PLATFORM=1 is set.
It was fixed by listed commit, checking super->orom was removed, but
also checking freesize. Freesize is not set for operations on RAID
volume with no size update, that's why it is not required to have
this value and always run autolayout_imsm().
Fix it by making autolayout_imsm() dependent on freesize.

Fixes: 46f192 ("imsm: fix first volume autolayout with IMSM_NO_PLATFORM")

Signed-off-by: Kinga Stefaniuk <kinga.stefaniuk@intel.com>